+++ docs/news.rst
@@ -1,8 +1,8 @@
 News
 ====
 
-Development (master)
---------------------
+3.2.1 (2019-03-19)
+------------------
 
 Minor features: